Address Feedback From PR Review
